460 Renames useSessionWithToken.ts to useWfoSession.ts
