+++ packages/orchestrator-ui-components/src/hooks/index.ts
@@ -10,5 +10,5 @@ export * from './useDataDisplayParams';
 export * from './useShowToastMessage';
 export * from './useStoredTableConfig';
 export * from './useWithOrchestratorTheme';
-export * from './useSessionWithToken';
+export * from './useWfoSession';
 export * from './useQueryWithFetch';
